more feedback & cleanup
